refactor: share the shimmer between the home and welcome marks
Extracted verbatim from StatusIcon, where it was private to one screen.
The welcome step sweeps the same glyph, and a second copy would drift in
band width or duration against the first.

The highlight becomes a parameter, defaulting to what StatusIcon already
passed, so the home glyph is unchanged. The status icon also grows to
match the welcome mark; its tint and its shimmer stay keyed to session
state, which is what that glyph is for.
